feat(cli): move CLI enter file for import types helper fn
