+++ packages/cli/bin/index.js
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 'use strict'
-require('../lib/index.js')
+require('../lib/main.js')
