+++ packages/cli/package.json
@@ -26,6 +26,7 @@
   ],
   "exports": {
     ".": "./lib/index.js",
+    "./main": "./lib/main.js",
     "./package.json": "./package.json"
   },
   "main": "lib/index.js",
