--- packages/cli/src/shared/types/config.ts
@@ -1,3 +1,0 @@
-export { UserConfig } from 'cz-git'
-export { CommitizenGitOptions } from 'cz-git'
-export { defineConfig, definePrompt } from 'cz-git'
